+++ sismic/code/python.py
@@ -109,22 +109,23 @@ class PythonEvaluator(Evaluator):
     Depending on the method that is called, the context can expose additional values:
 
     - Always:
-        - A *time* value that represents the current time exposed by the interpreter.
-        - An *active(name) -> bool* Boolean function that takes a state name and return *True* if and only
+        - A *time: float* value that represents the current time exposed by the interpreter.
+        - An *active(name: str) -> bool* Boolean function that takes a state name and return *True* if and only
           if this state is currently active, ie. it is in the active configuration of the ``Interpreter`` instance
           that makes use of this evaluator.
     - On code execution:
-        - A *send(name, **kwargs)* function that takes an event name and additional keyword parameters and fires
+        - A *send(name: str, **kwargs) -> None* function that takes an event name and additional keyword parameters and fires
           an internal event with.
-        - If the code is related to a transition, the *event* that fires the transition is exposed.
+        - If the code is related to a transition, the *event: Event* that fires the transition is exposed.
     - On code evaluation:
-        - If the code is related to a transition, the *event* that fires the transition is exposed.
+        - If the code is related to a transition, the *event: Event* that fires the transition is exposed.
     - On guard evaluation:
-        - An *after(sec) -> bool* Boolean function that returns *True* if and only if the source state
+        - An *after(sec: float) -> bool* Boolean function that returns *True* if and only if the source state
           was entered more than *sec* seconds ago. The time is evaluated according to Interpreter's clock.
-        - A *idle(sec) -> bool* Boolean function that returns *True* if and only if the source state
+        - A *idle(sec: float) -> bool* Boolean function that returns *True* if and only if the source state
           did not fire a transition for more than *sec* ago. The time is evaluated according to Interpreter's clock.
     - On postcondition or invariant:
+        - *after(sec: float) -> bool* and *idle(sec: float) -> bool*.
         - A variable *__old__* that has an attribute *x* for every *x* in the context when either the state
           was entered (if the condition involves a state) or the transition was processed (if the condition
           involves a transition). The value of *__old__.x* is a shallow copy of *x* at that time.
@@ -376,6 +377,10 @@ class PythonEvaluator(Evaluator):
 
         additional_context = {'event': event} if isinstance(obj, Transition) else {}  # type: Dict[str, Any]
         additional_context.update({'__old__': self.__memory.get(id(obj), None)})
+        additional_context.update({
+            'after': partial(self.__after, state_name),
+            'idle': partial(self.__idle, state_name),
+        })
 
         return filter(
             lambda c: not self._evaluate_code(c, context=context, additional_context=additional_context),
@@ -396,6 +401,10 @@ class PythonEvaluator(Evaluator):
 
         additional_context = {'event': event} if isinstance(obj, Transition) else {}  # type: Dict[str, Any]
         additional_context.update({'__old__': self.__memory.get(id(obj), None)})
+        additional_context.update({
+            'after': partial(self.__after, state_name),
+            'idle': partial(self.__idle, state_name),
+        })
 
         return filter(
             lambda c: not self._evaluate_code(c, context=context, additional_context=additional_context),